fixed circular imports between evaluate_inverse.py and evaluate_agent.py by moving
